Changed Versioning System and Commit Management
From this point in the commits on, the version system is changing to a more precise version.
vX.Y.Z :
X - Release version (will increment on full release and if any sequels are going to be still on this framework)
Y - Major version (will increment with any major updates, like for example new levels, new big mechanics being finished and so on)
Z - Minor version (will change with any increment in the project, not all commits will result in updated version, but any finished part will do so)

Several latest commits have rather larger and so from now on commits will try to be more elemental in changes and reasonable in size.
Git branching will be tested out more thoroughly, but if it will become a norm depends on how Godot will react to it.