Added trigger area object
An area2D that dynamically creates a colissionshape2D based.
Triggers rewritable methods when either the player or a NPC walks into the trigger area or leaves it (customizable with flags).
Designed to be inherited and methods rewritten to suit a specific need.
How many times the area can be triggered can be set. If set to a negative number it has no restriction.
Tracks how many times it was triggered. This persists through saving and loading (Persistence methods implemented).