Reworked dialogue variables to be dynamic
Previously all dialogue variables needed to be statically set into the given level to work. Now with two new methods they can be dynamically created and checked directly from the .dialogue script files.
check_variable method is to check if a given variable is set or what is its value.
set_variable method is to either set or rewrite a given variable.