Added a callable resource and more
Callable resource is an Editor way of setting up signal and method calls without the need for creating methods specifically for calling those methods and signals.
It is a resource that can be attached in a list to any node and then. Then in the editor it can be setup with a nodepath or an autoload script name, method or signal name and parameters that should be sent with the call. When ever needed the node then can go through the list of callables and execute each one without the need of any static setups inside the code of the node.
For example now instead of an interactable object having to have a method inside itself that can call for a dialogue to be triggered and having to have a dialogue resource exported, it just has an exported list of callable resources. One those entries now can be a callable resource that triggers the dialogue and that also holds the dialogue resource.
This now allows much quicker and dynamic creation of content and interactions between object within the game.
Updated interactable objects and npcs with this new callable version.