Added a timeline selection view and more
Added a timeline selection mechanics for selecting which timeline the player will shift into and can foresee into.
Reworked the timeline controller to work with this mechanic.
Reworked some other scripts that work with timelines.
Added translation for this mechanic.
Added textures for this mechanic, symbols for left and right mouse click.